fix: move GHContext from gesture-handler to native-stack (#2017)
## Description

Unfortunately, because of changing the way how we export the components,
I didn't saw the wrong export of GHContext in gesture-handler's
`index.tsx` file. To match the changes from the PR with full screen
swipe (#1913), I've created the directory in a wrong place. This PR
fixes this export by moving the context to native-stack and removing the
export from the gesture-handler/index.tsx.
I'm also removing this export, because right now it's possible to import
GHContext from `react-native-screens` and also from
`react-native-screens/gesture-handler` which is wrong.

## Changes

- Changed the export of GHContext in index.tsx
- Moved GHContext from gesture-handler to native-stack

## Test code and steps to reproduce

Try to build and run TestsExample - see if the bundling state passes
correctly. Then, try to import GestureDetectorProvider and wrap whole
navigator in it.

## Checklist

- [X] Included code example that can be used to test this change
- [x] Ensured that CI passes